Moved repository
Moved this project to a new repository. It can't be considered a fork
of the old project as it can't be merged into it due to being so different.
It needs its own releases and so on.

Added npm badge to readme.
Update package.dist.json.
I removed the scripts and updated all git repository links.